Add files via upload
Equity management is the process of creating and managing owners in a company. They can be broadly categorized into ‘active’ and ‘passive’ strategies. A common passive strategy is to have an index fund whose portfolio mirrors the composition and performance of a specific index, such as the S&P 500. This can be easily done by purchasing all the stocks in the index with the corresponding weights. However, this quickly becomes expensive due to rebalancing costs from trading. 

Thus, this project attempts to create an index fund that best tracks the NASDAQ-100 with a fewer number of stocks compared to the index. This can be achieved by formulating an integer program to pick exactly m out of the n stocks for the fund. We will also use a linear program to identify the optimal weights for these stocks such that the absolute differences between the portfolio returns and the index returns are minimized.